refactor: inital refactor all
