+++ client/src/components/ProtectedRoute.tsx
