+++ client/src/router.tsx
@@ -1,20 +1,35 @@
 import { Navigate, Route, Routes } from "react-router-dom"
 
-import AgentRoutes from "./apps/Agent/routes"
-import TuristRoutes from "./apps/Turist/routes"
+import Home from "./pages/HomeTurist/Home";
+import SignIn from "./pages/SignIn/SignIn";
+import SignUp from "./pages/SignUp/SignUp";
+import ProtectedRoute from "./components/ProtectedRoute";
+
+import Home from "./pages/Home/Home";
+import CreateTouristLocation from "./pages/CreateTuristLocation/CreateTouristLocation";
+import Profile  from "./pages/Profile/Profile";
 
 const AppRoutes = () => {
     return (
         <Routes>
-            <Route index element={<Navigate to={"/home"} />} />
-            {/* 
+            <Route index element={<Navigate to="/signin" />} />
 
-            <Route path={"/account/login"} element={<SignIn />} />
-            <Route path={"/account/register"} element={<SignUp />} />
-             */}
+            <Route
+                path="/home"
+                element={
+                    <ProtectedRoute>
+                        <Home />
+                    </ProtectedRoute>
+                }
+            />
 
-            <Route path={"/*"} element={<TuristRoutes />} />
-            <Route path={"/agent/*"} element={<AgentRoutes />} />
+            <Route path="/signin" element={<SignIn />} />
+            <Route path="/signup" element={<SignUp />} />
+            <Route index element={<Navigate to={"/home-agent"} />} />
+
+            <Route path="/home-agent" element={<Home />} />
+            <Route path="/create/tourist-place" element={<CreateTouristLocation />} />
+            <Route path="/profile" element={<Profile />} />
         </Routes>
     )
 }
